+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ---
 
+### 0.0.2 : combined scrollers class into Sliders() 
 ### 0.0.1 : separated from bldrs_app
 
 ---
@@ -8,5 +9,6 @@
 
 * Scroller(); // ScrollBar Widget
 * Animators(); // Animation helpers class
+* Sliders(); // Scrolling, Swiping And Sliding helpers class
 
 ---